NXT-489 Relocated base testimonial component to shared folder
